refactor: 컴포넌트 리네이밍
