+++ src/components/book/BookSpecs.css.ts
